Rename jellyfinactions.js to jellyfinActions.ts
